feat(ai): filter AI tickets
This commit ensures that tickets redeemed for AI tickets are shown with
a different message and colour in Discord and Twitter. The current
implementation is a quick POC in which messages are filtered based on
the sender being one of the known AI broadcasters. I will add a
`jobType` field to the `auxData` field when we move into the **beta**
phase of the subnet, and any broadcaster can join.